nearly complete coverage of main
refactor a bit
